Add Python sample script and C test file
image_processing.py didn't get copied over in
6d8b8f391712d9d2cb225f0616cc3e628e47a444, so add it now.

py_test.c was used to test the development of the Python interface, and is
added here for reference. To embed the Python interpreter and link against
Python, the Python development headers and libraries are needed. This can be
installed with "sudo apt install python-dev". To compile and run:
1. gcc -I /usr/include/python3.11 -c py_test.c -o py_test.o
2. gcc -I /usr/include/python3.11 -c py_interface.c -o py_interface.o
3. gcc -c app_log.c -o app_log.o
4. gcc py_test.o py_interface.o app_log.o -lpython3.11 -o program
5. ./program